Use double precision for profile calculation
Single precision math results in sizeable errors
in Chi^2 in some cases, and differences between
platforms. Since double precision math is used
everywhere else in IMP, and does not appear to have
a noticeable effect on FoXS CPU or memory usage in our
testing, switch IMP.saxs to use double precision Eigen
vectors and matrices.

+++ modules/saxs/src/ChiFreeScore.cpp
@@ -40,8 +40,8 @@ double ChiFreeScore::compute_score(const Profile* exp_profile,
 
   IMP_NEW(Profile, exp_profile_selection, ());
   IMP_NEW(Profile, model_profile_selection, ());
-  Eigen::VectorXf qs(ns_), errors(ns_);
-  Eigen::VectorXf exp_intensities(ns_), model_intensities(ns_);
+  Eigen::VectorXd qs(ns_), errors(ns_);
+  Eigen::VectorXd exp_intensities(ns_), model_intensities(ns_);
 
   // repeat K_ times
   for (unsigned int k = 0; k < K_; k++) {
